Address review comments for PR #26
- Fix log units and doc comments in device code
- Clean up unused imports and globals in usb_cdc_receiver
- Fix sequence number handling and type consistency (u32 -> u16)
- Implement proper device counting and stats updates
- Use constants for protocol header length in Python handler
- Add unit tests for nested frame unpacking
- Remove restrictive rust-toolchain.toml

+++ server/usb_cdc_receiver/src/streaming/controller.rs
@@ -262,7 +262,7 @@ impl StreamingController {
         // データフレームタイプを決定（実際のフレームタイプに基づく）
         let acked_message_type = MessageType::DataFrame; // 現在はすべてデータフレームとして扱う
         
-        let ack = AckMessage::new(frame.sequence as u16, acked_message_type, status);
+        let ack = AckMessage::new(frame.sequence, acked_message_type, status);
         let ack_data = ack.serialize();
         
         match self.esp_now_sender.send_data(mac_address, &ack_data) {
